requirements.txt
akshare
pandas
numpy
matplotlib
requests
yfinance
openpyxl
pillow
